+++ Expanded Cheat Menu/Code/Menus/DebugMenu.lua
@@ -3,7 +3,6 @@
 local Concat = ChoGGi.ComFuncs.Concat
 local S = ChoGGi.Strings
 local Actions = ChoGGi.Temp.Actions
-local blacklist = ChoGGi.blacklist
 
 Actions[#Actions+1] = {
   ActionMenubar = "Debug",
@@ -72,17 +71,17 @@ Actions[#Actions+1] = {
   OnAction = ChoGGi.MenuFuncs.ShowAnimDebug_Toggle,
 }
 
---~ Actions[#Actions+1] = {
---~   ActionMenubar = "Debug",
---~   ActionName = S[302535920000471--[[Object Manipulator--]]],
---~   ActionId = "Debug.Object Manipulator",
---~   ActionIcon = "CommonAssets/UI/Menu/SaveMapEntityList.tga",
---~   RolloverText = S[302535920000472--[[Manipulate objects (selected or under mouse cursor)--]]],
---~   OnAction = function()
---~     ChoGGi.ComFuncs.OpenInObjectManipulator()
---~   end,
---~   ActionShortcut = ChoGGi.UserSettings.KeyBindings.OpenInObjectManipulator,
---~ }
+Actions[#Actions+1] = {
+  ActionMenubar = "Debug",
+  ActionName = S[302535920000471--[[Object Manipulator--]]],
+  ActionId = "Debug.Object Manipulator",
+  ActionIcon = "CommonAssets/UI/Menu/SaveMapEntityList.tga",
+  RolloverText = S[302535920000472--[[Manipulate objects (selected or under mouse cursor)--]]],
+  OnAction = function()
+    ChoGGi.ComFuncs.OpenInObjectManipulatorDlg()
+  end,
+  ActionShortcut = ChoGGi.UserSettings.KeyBindings.OpenInObjectManipulator,
+}
 
 Actions[#Actions+1] = {
   ActionMenubar = "Debug",